// src/api/models/OrganizationModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizationModel = {
  name: string;
  description: string;
  logoUrl?: string | null;
  dateUpdated?: string | null;
  websiteUrl?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
};